refactor exporter
